Init commit - Label generator consolidation
Initial work for consolidating the label generator into the smile-server codebase
