+++ server/src/main/java/org/mskcc/smile/SmileApp.java
@@ -4,11 +4,11 @@ import java.util.concurrent.CountDownLatch;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.mskcc.cmo.messaging.Gateway;
-import org.mskcc.smile.service.ClinicalMessageHandlingService;
-import org.mskcc.smile.service.CorrectCmoPatientHandlingService;
-import org.mskcc.smile.service.RequestReplyHandlingService;
-import org.mskcc.smile.service.ResearchMessageHandlingService;
-import org.mskcc.smile.service.TempoMessageHandlingService;
+import org.mskcc.smile.service.msg.ClinicalMessageHandlingService;
+import org.mskcc.smile.service.msg.CorrectCmoPatientHandlingService;
+import org.mskcc.smile.service.msg.RequestReplyHandlingService;
+import org.mskcc.smile.service.msg.ResearchMessageHandlingService;
+import org.mskcc.smile.service.msg.TempoMessageHandlingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
