+++ service/src/main/java/org/mskcc/smile/service/msg/CorrectCmoPatientHandlingService.java
@@ -1,4 +1,4 @@
-package org.mskcc.smile.service;
+package org.mskcc.smile.service.msg;
 
 import java.util.Map;
 import org.mskcc.cmo.messaging.Gateway;
