+++ service/src/main/java/org/mskcc/smile/service/msg/impl/RequestReplyHandlingServiceImpl.java
@@ -1,4 +1,4 @@
-package org.mskcc.smile.service.impl;
+package org.mskcc.smile.service.msg.impl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.nats.client.Message;
@@ -19,7 +19,7 @@ import org.mskcc.cmo.messaging.MessageConsumer;
 import org.mskcc.smile.model.SampleMetadata;
 import org.mskcc.smile.model.SmileSample;
 import org.mskcc.smile.service.CrdbMappingService;
-import org.mskcc.smile.service.RequestReplyHandlingService;
+import org.mskcc.smile.service.msg.RequestReplyHandlingService;
 import org.mskcc.smile.service.SmileSampleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
