+++ service/src/test/java/org/mskcc/smile/service/SmileTestApp.java
@@ -5,16 +5,16 @@ import org.mskcc.cmo.messaging.utils.SSLUtils;
 import org.mskcc.smile.commons.JsonComparator;
 import org.mskcc.smile.commons.impl.JsonComparatorImpl;
 import org.mskcc.smile.persistence.jpa.CrdbRepository;
-import org.mskcc.smile.service.impl.ClinicalMessageHandlingServiceImpl;
+import org.mskcc.smile.service.msg.impl.ClinicalMessageHandlingServiceImpl;
 import org.mskcc.smile.service.impl.CohortCompleteServiceImpl;
-import org.mskcc.smile.service.impl.CorrectCmoPatientHandlingServiceImpl;
+import org.mskcc.smile.service.msg.impl.CorrectCmoPatientHandlingServiceImpl;
 import org.mskcc.smile.service.impl.CrdbMappingServiceImpl;
 import org.mskcc.smile.service.impl.PatientServiceImpl;
-import org.mskcc.smile.service.impl.RequestReplyHandlingServiceImpl;
+import org.mskcc.smile.service.msg.impl.RequestReplyHandlingServiceImpl;
 import org.mskcc.smile.service.impl.RequestServiceImpl;
-import org.mskcc.smile.service.impl.ResearchMessageHandlingServiceImpl;
+import org.mskcc.smile.service.msg.impl.ResearchMessageHandlingServiceImpl;
 import org.mskcc.smile.service.impl.SampleServiceImpl;
-import org.mskcc.smile.service.impl.TempoMessageHandlingServiceImpl;
+import org.mskcc.smile.service.msg.impl.TempoMessageHandlingServiceImpl;
 import org.mskcc.smile.service.impl.TempoServiceImpl;
 import org.mskcc.smile.service.util.RequestStatusLogger;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
